fix(ui): comfirm modal for removing secrets/networks/configs [EE-4211]

+++ app/docker/views/networks/networksController.js
@@ -11,26 +11,32 @@ angular.module('portainer.docker').controller('NetworksController', [
   'HttpRequestHelper',
   'endpoint',
   'AgentService',
-  function ($q, $scope, $state, NetworkService, Notifications, HttpRequestHelper, endpoint, AgentService) {
+  'ModalService',
+  function ($q, $scope, $state, NetworkService, Notifications, HttpRequestHelper, endpoint, AgentService, ModalService) {
     $scope.removeAction = function (selectedItems) {
-      var actionCount = selectedItems.length;
-      angular.forEach(selectedItems, function (network) {
-        HttpRequestHelper.setPortainerAgentTargetHeader(network.NodeName);
-        NetworkService.remove(network.Id)
-          .then(function success() {
-            Notifications.success('Network successfully removed', network.Name);
-            var index = $scope.networks.indexOf(network);
-            $scope.networks.splice(index, 1);
-          })
-          .catch(function error(err) {
-            Notifications.error('Failure', err, 'Unable to remove network');
-          })
-          .finally(function final() {
-            --actionCount;
-            if (actionCount === 0) {
-              $state.reload();
-            }
-          });
+      ModalService.confirmDeletion('Do you want to remove the selected network(s)?', (confirmed) => {
+        if (!confirmed) {
+          return;
+        }
+        var actionCount = selectedItems.length;
+        angular.forEach(selectedItems, function (network) {
+          HttpRequestHelper.setPortainerAgentTargetHeader(network.NodeName);
+          NetworkService.remove(network.Id)
+            .then(function success() {
+              Notifications.success('Network successfully removed', network.Name);
+              var index = $scope.networks.indexOf(network);
+              $scope.networks.splice(index, 1);
+            })
+            .catch(function error(err) {
+              Notifications.error('Failure', err, 'Unable to remove network');
+            })
+            .finally(function final() {
+              --actionCount;
+              if (actionCount === 0) {
+                $state.reload();
+              }
+            });
+        });
       });
     };
