+++ app/docker/views/secrets/secretsController.js
@@ -4,26 +4,31 @@ angular.module('portainer.docker').controller('SecretsController', [
   'SecretService',
   'Notifications',
   function ($scope, $state, SecretService, Notifications) {
-    $scope.removeAction = function (selectedItems) {
-      var actionCount = selectedItems.length;
-      angular.forEach(selectedItems, function (secret) {
-        SecretService.remove(secret.Id)
-          .then(function success() {
-            Notifications.success('Secret successfully removed', secret.Name);
-            var index = $scope.secrets.indexOf(secret);
-            $scope.secrets.splice(index, 1);
-          })
-          .catch(function error(err) {
-            Notifications.error('Failure', err, 'Unable to remove secret');
-          })
-          .finally(function final() {
-            --actionCount;
-            if (actionCount === 0) {
-              $state.reload();
-            }
-          });
-      });
-    };
+    this.ModalService.confirmDeletion('Do you want to remove the selected secret(s)?', (confirmed) => {
+      if (!confirmed) {
+        return;
+      }
+      $scope.removeAction = function (selectedItems) {
+        var actionCount = selectedItems.length;
+        angular.forEach(selectedItems, function (secret) {
+          SecretService.remove(secret.Id)
+            .then(function success() {
+              Notifications.success('Secret successfully removed', secret.Name);
+              var index = $scope.secrets.indexOf(secret);
+              $scope.secrets.splice(index, 1);
+            })
+            .catch(function error(err) {
+              Notifications.error('Failure', err, 'Unable to remove secret');
+            })
+            .finally(function final() {
+              --actionCount;
+              if (actionCount === 0) {
+                $state.reload();
+              }
+            });
+        });
+      };
+    });
 
     $scope.getSecrets = getSecrets;
